day 18: some neat changes
- don't pass the full list of corruptions and a byte count, but
  instead just pass the slice of corruptions up to the byte count!
- this also simplifies working with the list of corruptions in
  the solve function as well
- in part 2 use an empty for loop that walks backwards from the
  full list of corruptions as long as there is no path to the exit
- that loop uses the result of the solve() function to determine
  when to stop decrementing the counter
- declaring the variable with var instead of let hoists it to the
  function scope of part2(), so variable c is accessible outside
  of the loop. Back to the older days of JS :)

That way the definition of the head of the loop does all the heavy
lifting in part 2: declaring the variable and keep decreasing it
until a path is found.

You could also move the definition of the corruption variable into
the for loop, but while this would save another line of code, it
would also make for a very long and busy line and I prefer to
keep variables that never change defined as constants.